src/accounts/contacts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.2.0
    # via -r requirements.in
bcrypt==4.1.2
    # via -r requirements.in
bleach==6.1.0
    # via -r requirements.in
blinker==1.7.0
    # via flask
cachetools==5.3.3
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   -r requirements.in
    #   requests
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r requirements.in
    #   flask
coverage[toml]==7.4.4
    # via pytest-cov
cryptography==42.0.5
    # via -r requirements.in
flask==3.0.2
    # via -r requirements.in
grpcio==1.62.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==6.11.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.3
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
more-itertools==10.2.0
    # via -r requirements.in
packaging==23.2
    # via
    #   -r requirements.in
    #   gunicorn
    #   pytest
pluggy==1.4.0
    # via
    #   -r requirements.in
    #   pytest
protobuf==4.25.3
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
py==1.11.0
    # via -r requirements.in
pyasn1==0.5.1
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via -r requirements.in
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pycryptodome==3.20.0
    # via -r requirements.in
pyjwt==2.8.0
    # via -r requirements.in
pyparsing==3.1.2
    # via -r requirements.in
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements.in
pytz==2023.4
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
rsa==4.9
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   bleach
sqlalchemy==1.4.52
    # via -r requirements.in
urllib3==2.2.1
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.2.13
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   -r requirements.in
    #   bleach
werkzeug==3.0.1
    # via flask
wrapt==1.16.0
    # via -r requirements.in
zipp==3.17.0
    # via
    #   -r requirements.in
    #   importlib-metadata
